Move ShardId to model
